feat: Implement UserService methods for granting/revoking premium role and prevent admin self-deletion

+++ src/main/java/com/montelzek/moneytrack/service/UserServiceImpl.java
@@ -1,22 +1,26 @@
 package com.montelzek.moneytrack.service;
 
+import com.montelzek.moneytrack.model.Role;
 import com.montelzek.moneytrack.model.User;
 import com.montelzek.moneytrack.repository.UserRepository;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 
 @Service
 public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
 
-    public UserServiceImpl(UserRepository userRepository) {
+    private final RoleService roleService;
+
+    public UserServiceImpl(UserRepository userRepository, RoleService roleService) {
         this.userRepository = userRepository;
+        this.roleService = roleService;
     }
 
     @Override
@@ -40,7 +44,14 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
+    @Transactional
     public void deleteById(Long id) {
+
+        Long currentUserId = getCurrentUserId();
+        if (currentUserId != null && currentUserId.equals(id)) {
+            throw new IllegalArgumentException("Admin cannot delete themselves.");
+        }
+
         userRepository.deleteById(id);
     }
 
@@ -52,5 +63,41 @@ public class UserServiceImpl implements UserService {
     @Override
     public Boolean existsByEmail(String email) {
         return userRepository.existsByEmail(email);
+    }
+
+    @Override
+    @Transactional
+    public void grantPremiumRole(Long userId) {
+        updateUserRole(userId, true);
+    }
+
+    @Override
+    @Transactional
+    public void revokePremiumRole(Long userId) {
+        updateUserRole(userId, false);
+    }
+
+    private Role getRole(Role.ERole roleName) {
+        return roleService.findByName(roleName)
+                .orElseThrow(() -> new RuntimeException("Role " + roleName + " not found."));
+    }
+
+    private void updateUserRole(Long userId, boolean addRole) {
+
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new UsernameNotFoundException("User not found with id: " + userId));
+        Set<Role> roles = user.getRoles();
+
+        if (roles.contains(getRole(Role.ERole.ROLE_ADMIN))) {
+            return;
+        }
+
+        Role targetRole = getRole(Role.ERole.ROLE_PREMIUM);
+
+        boolean changed = addRole ? roles.add(targetRole) : roles.remove(targetRole);
+
+        if (changed) {
+            userRepository.save(user);
+        }
     }
 }
